+++ micromouse_source/micromouse_source/config/config.h
@@ -1,9 +1,10 @@
 #ifndef MICROMOUSE_CONFIG_H
 #define MICROMOUSE_CONFIG_H
+#include <cstdint>
 
 namespace Config
 {
-	const extern int PCA9685_DEVICE_ID;
+	const extern uint_fast8_t PCA9685_DEVICE_ID;
 }
 
 #endif // MICROMOUSE_CONFIG_H
